+++ ant_gen/__init__.py
@@ -1,5 +1,5 @@
 '''Top Level Package for antenna generator'''
 
-__version__ = '0.3.1'
+__version__ = '0.3.2'
 __author__ = 'Joyen Benitto'
 __mail__ = 'joyen.benitto12@gmail.com'