+++ ant_gen/main.py
@@ -1,6 +1,7 @@
 import ant_gen.generator as gen
 import ant_gen.utils as utils
 from ant_gen.__init__ import __version__
+from ant_gen.__init__ import __name__
 import os
 import click
 
@@ -8,7 +9,7 @@ import click
 @click.group()
 @click.version_option(version=__version__)
 def cli():
-	'''Command Line Interface for antenna generator'''
+	f'''Command Line Interface for antenna generator'''
 
 @click.version_option(version=__version__)
 # CLI option 'log'.
@@ -34,7 +35,7 @@ def generate(ant, build_dir):
     if ant == "msp":
         utils.clean_dir(build_dir)
         os.system(f"mkdir {build_dir}")
-        gen.generator(build_dir) 
+        gen.generator_msp(build_dir) 
     else:
         utils.clean_dir(build_dir)
-        print("Provide a valid antenna")
+        print("Provide a valid antenna !")