+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # CHANGELOG
 
+## [0.3.2] - 2023-07-26
+### Added:
+- Added L slot optimization
+- Added slot optimization
+
 ## [0.3.1] - 2023-07-25
 ### Added:
+- Added Cut-out optimization
+- Added space for experiment setup
+- Added click-up options
+- Added Subtract functionality
 - Added click commands
 
 ## [0.3.0] - 2023-07-22
@@ -11,12 +20,12 @@
 - Added pass/fail yaml examples 
 - Added unite functionality
 - Added subtract functionality
----
+
 ## [0.2.0] - 2023-07-20
 ### Added:
 - Schema structure 
 - Multiple patch generation
----
+
 ## [0.1.0] - 2023-07-17
 ### Added:
 - Initialized the Antenna generator